requirements.txt
PyYAML==5.4.1
boto3==1.18.15
botocore==1.21.15
pyparsing==2.4.7
requests==2.26.0
retrying==1.3.3
Jinja2==3.0.3
pytest-shutil==1.7.0
